+++ backend/canisters/community/impl/src/updates/accept_p2p_swap.rs
@@ -26,7 +26,7 @@ async fn accept_p2p_swap_impl(args: Args) -> Response {
         Err(response) => return Error(response),
     };
 
-    let result = match user_canister_c2c_client::c2c_accept_p2p_swap(user_id.into(), &c2c_args).await {
+    let result = match user_canister_c2c_client::c2c_accept_p2p_swap(user_id.canister_id(), &c2c_args).await {
         Ok(user_canister::c2c_accept_p2p_swap::Response::Success(transaction_index)) => {
             NotifyEscrowCanisterOfDepositJob::run(
                 user_id,