+++ backend/canisters/group_index/impl/src/jobs/push_group_deleted_notifications.rs
@@ -55,7 +55,7 @@ async fn push_notifications(notifications: Vec<(UserId, DeletedGroupInfoInternal
 async fn push_notification(user_id: UserId, deleted_group: DeletedGroupInfoInternal) {
     let args = user_canister::c2c_notify_group_deleted::Args { deleted_group };
 
-    if user_canister_c2c_client::c2c_notify_group_deleted(user_id.into(), &args)
+    if user_canister_c2c_client::c2c_notify_group_deleted(user_id.canister_id(), &args)
         .await
         .is_err()
     {